Initiate protocol for Bobert dating algorithm 💘
Bobert is a legend in his own unique way. We named our band in high school after his last name, and he was a great guitarist in the band. He knew my sister, her friends, and many of the people that came to be my close confidantes in high school and college. He went to college early, and had a great education. He worked his way up from an intern to become the leader of a team at a social media tech company, and I always turn to him when I have a question about code or systems design. He is fit and healthy, and has a great sense of humor, sarcastic and witty, and he loves finding trendy spots to eat; he has a knack for finding these spots which are just, extraordinary, and uncommon.

He's close to his family. He cares about his parents, he cares about his sibling, and he knows that's he's the best in the west at AI research. He wants to meet someone from another place, and build a family. He's a good guy, ring that bell, bell.